sync: Extract out pipeline instance runner code
Change-type: patch
