+++ lib/sync/pipeline.js
@@ -5,51 +5,14 @@
  */
 
 const _ = require('lodash')
-const request = require('request')
 const Bluebird = require('bluebird')
 const jsone = require('json-e')
 const errors = require('./errors')
-
-// TODO: Make this function understand OAuth, and allow
-// it to re-fresh and update OAuth tokens from user cards
-// transparently to the integrations.
-const httpRequest = async (options) => {
-	return new Bluebird((resolve, reject) => {
-		request(options, (error, response, body) => {
-			if (error) {
-				return reject(error)
-			}
-
-			return resolve({
-				code: response.statusCode,
-				body
-			})
-		})
-	})
-}
+const instance = require('./instance')
 
 const runIntegration = async (integration, options, fn, card) => {
-	// eslint-disable-next-line new-cap
-	const instance = new integration({
-		// Notice that integrations don't have access at all
-		// to functions that insert/upsert to the data store
-		context: {
-			log: options.context.log,
-			getElementBySlug: options.context.getElementBySlug,
-			getElementById: options.context.getElementById,
-			getActorId: options.context.getActorId,
-			getElementByMirrorId: options.context.getElementByMirrorId,
-			request: httpRequest
-		},
-
-		errors,
-		token: options.token
-	})
-
-	await instance.initialize()
-
-	try {
-		const sequence = await instance[fn](card, {
+	return instance.run(integration, options.token, async (integrationInstance) => {
+		const sequence = await integrationInstance[fn](card, {
 			actor: options.actor
 		})
 
@@ -58,16 +21,13 @@ const runIntegration = async (integration, options, fn, card) => {
 			sequence
 		})
 
-		const result = await exports.importCards(options.context, sequence, {
+		return exports.importCards(options.context, sequence, {
 			origin: card
 		})
-
-		await instance.destroy()
-		return result
-	} catch (error) {
-		await instance.destroy()
-		throw error
-	}
+	}, {
+		actor: options.actor,
+		context: options.context
+	})
 }
 
 /**
